Split Lexer and Lexinator into seperate files
